Separate interface and Implementation.
